backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-limiter==3.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
gunicorn==21.2.0
PyMuPDF==1.23.8
reportlab==4.0.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
APScheduler==3.10.4

